#205 - Finalize restructure of components directory tree
