+++ src/components/whole-app/login/login.tsx
@@ -8,7 +8,7 @@ import { useContext } from 'react';
 import { useHistory } from 'react-router';
 import { Form, InputGroup, FormControl, Button } from 'react-bootstrap';
 import { UserLogInContext } from '../../app/app-context/app-context';
-import { routes } from '../../shared/routes';
+import { routes } from '../../../shared/routes';
 import styles from './login.module.css';
 
 /**
